tryna add admin and post auth

+++ app/api/event/delete/route.ts
@@ -3,6 +3,12 @@ import prisma from '@/lib/prisma';
 export const POST = async (req: NextRequest) => {
   const eventId = await req.json();
   try {
+    const user=await prisma.user.findUnique({
+      where:{
+        clerkId:eventId.userId
+      }
+    })
+    if(!user||!user.canCreateEvents)return NextResponse.json({msg:"Auth me issue sa aagya thoda"},{status:403})
     const delEvent= await prisma.event.delete({
       where:{
         eventId:eventId.eventId
